dual funding: update language around change outputs
Couple of changes. First, make the inclusion of a change output
optional on a funding transaction. Second, we should be measuring
the inclusion of the change output against the `dust_limit_satoshi`,
not zero.

Finally, update language to indicate that any fees that eat through
the 'change' amount (if included) will be deduced from the sum of
the parties' `funding_satoshis`, e.g. the accepter will pay for the
remainder of the fees in the case that the `funding_satoshis` of
the opener are completely consumed by fee costs.